# Server/src/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "TODO Task API",
        "description": "Node.js Application API"
    },
    "host": "https://va811ai7l5.execute-api.us-east-1.amazonaws.com/dev",
    "basePath": "/api",
    "tags": [
        {
            "name": "users",
            "description": "API for users in the system"
        },
        {
            "name": "tasks",
            "description": "API for tasks in the system"
        }
    ],
    "schemes": [
        "http",
        "https"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "securityDefinitions": {
        "APIKeyHeader": {
            "type": "apiKey",
            "in": "header",
            "name": "authorization"
        }
    },
    "security": [
        {
            "APIKeyHeader": []
        }
    ],
    "paths": {
        "/users": {
            "get": {
                "tags": [
                    "users"
                ],
                "summary": "Get all users in system",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/users/me": {
            "get": {
                "tags": [
                    "users"
                ],
                "summary": "Get user profile",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/users/setup-notification": {
            "post": {
                "tags": [
                    "users"
                ],
                "summary": "Setup user notification",
                "parameters": [
                    {
                        "name": "notificationToken",
                        "in": "body",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/users/push-notifications": {
            "put": {
                "tags": [
                    "users"
                ],
                "summary": "Setup user notification",
                "parameters": [
                    {
                        "name": "notificationToken",
                        "in": "body",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/tasks/add": {
            "post": {
                "tags": [
                    "tasks"
                ],
                "summary": "Add task",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Task object that needs to be added",
                        "schema": {
                            "$ref": "#/definitions/Task"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/tasks/update": {
            "put": {
                "tags": [
                    "tasks"
                ],
                "summary": "Update task",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Task object that needs to be updated",
                        "schema": {
                            "$ref": "#/definitions/Task"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/tasks/delete/{taskId}": {
            "delete": {
                "tags": [
                    "tasks"
                ],
                "summary": "Delete task",
                "parameters": [
                    {
                        "name": "taskId",
                        "in": "path",
                        "description": "Id of task to delete",
                        "required": true,
                        "type": "string",
                        "format": "ObjectId"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/tasks": {
            "get": {
                "tags": [
                    "tasks"
                ],
                "summary": "Get tasks",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/tasks/assigned": {
            "get": {
                "tags": [
                    "tasks"
                ],
                "summary": "Get Task List By User",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "definitions": {
        "Task": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "format": "ObjectId"
                },
                "title": {
                    "type": "string",
                    "required": true
                },
                "description": {
                    "type": "string"
                },
                "assignedTo": {
                    "type": "string",
                    "format": "ObjectId"
                }
            }
        }
    }
}